Move WaterDebugSnapshot to Core/Managers/World/Debug
