+++ AAEmu.Game/Core/Managers/World/WorldManager.cs
@@ -8,6 +8,7 @@ using AAEmu.Commons.IO;
 using AAEmu.Commons.Utils;
 using AAEmu.Game.Core.Managers.Id;
 using AAEmu.Game.Core.Managers;
+using AAEmu.Game.Core.Managers.World.Debug;
 using AAEmu.Game.Core.Network.Game;
 using AAEmu.Game.Core.Packets.G2C;
 using AAEmu.Game.IO;
